Fix next/dynamic
